rename gstream to speech
