+++ src/speech.rs
